+++ plugins/jira/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/jira
 
+## 1.2.42
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.21.9
+
 ## 1.2.41
 
 ### Patch Changes