+++ testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # testing
 
+## 0.1.45
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.21.9
+  - @bs-plugins/jira@1.2.42
+  - @bs-plugins/template@0.2.41
+
 ## 0.1.44
 
 ### Patch Changes